+++ code/modules/reagents/reagent_containers/syringes.dm
@@ -243,10 +243,12 @@
 			if((user != target) && H.check_shields(7, src, user, "\the [src]"))
 				return
 
-			if (target != user && H.getarmor(target_zone, "melee") > 5 && prob(50))
-				for(var/mob/O in viewers(world.view, user))
-					O.show_message(text("\red <B>[user] tries to stab [target] in \the [hit_area] with [src.name], but the attack is deflected by armor!</B>"), 1)
-				user.remove_from_mob(src)
+			if(target != user && H.getarmor(target_zone, "melee") > 5 && prob(50))
+				if(!user.unEquip(src))
+					return
+				user.show_message(
+					SPAN_DANGER("[user] tries to stab [target] in \the [hit_area] with [src.name], but the attack is deflected by armor!")
+				)
 				qdel(src)
 
 				user.attack_log += "\[[time_stamp()]\]<font color='red'> Attacked [target.name] ([target.ckey]) with \the [src] (INTENT: HARM).</font>"
